Week 3 - EXE
I went with a short labyrinth, a lot of the time I spent was figuring out the assets, how to change the landscape, and making the other person in the maze chase you. A couple thing I want to change, walk speed, make an ambiance sound in the background, add a health system, maybe traps etc. I think Unreal has a lot of potential and I'm already excited to use it!